Add basic README description, fixed a bug in the read of the Segment module
The segment module had a bugged edge case for which when attempting to
read a record in an offset postion above the latest written in the index
file, the record returned was wrongly calculated, instead we want to
consume the entire remaining bytes from the log file until we find the
requested offset OR we exhaust the buffer (e.g. the requested offset
is higher than the actual last offset inseted).